fix imports, update tests
