+++ tests/dkg_utils.py
@@ -3,7 +3,7 @@
 from sgx import SgxClient
 from sgx.sgx_rpc_handler import SgxServerError
 
-from tests.conftest import TestSettings
+from tests.fixtures.settings import TestSettings
 from tools.constants import SGX_CERTIFICATES_FOLDER
 
 INSECURE_PRIVATE_KEY = 'f253bad7b1f62b8ff60bbf451cf2e8e9ebb5d6e9bff450c55b8d5504b8c63d3'
