+++ tests/schains/ssl_test.py
@@ -2,7 +2,7 @@ import os
 from datetime import datetime
 
 from core.chain.ssl import get_ssl_filepath, get_ssl_files_change_date
-from tools.configs import NODE_DATA_PATH
+from tools.constants import NODE_DATA_PATH
 
 
 def test_get_ssl_filepath(cert_key_pair):
