+++ tools/constants/schains.py
@@ -19,11 +19,10 @@
 
 import os
 
-from tools.constants import CONFIG_FOLDER, NODE_DATA_PATH, NODE_DATA_PATH_HOST
+from tools.constants import CONFIG_FOLDER, NODE_DATA_PATH
 
 SCHAINS_DIR_NAME = 'schains'
 SCHAINS_DIR_PATH = os.path.join(NODE_DATA_PATH, SCHAINS_DIR_NAME)
-SCHAINS_DIR_PATH_HOST = os.path.join(NODE_DATA_PATH_HOST, SCHAINS_DIR_NAME)
 
 BASE_SCHAIN_CONFIG_FILENAME = 'schain_base_config.json'
 BASE_SCHAIN_CONFIG_FILEPATH = os.path.join(CONFIG_FOLDER, BASE_SCHAIN_CONFIG_FILENAME)
